Remove Ping from New* methods
Previously calling `New` or `NewInsecure` would trigger a call to `Ping`
without giving the caller the ability to set the logger or configure
client behaviour.

This removes the call to `Ping`, which lets the caller do whatever they
want to modify the default behaviour before they make any network requests.